# master/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: birdman master API
  version: v1
  description: |-
    Public REST API of birdman-master: matchmaking, fleet control, deploys and observability.

    This document is GENERATED from the master's route table (master/internal/httpapi/routes.go)
    and cannot drift from the router: the router is registered from that same table.

    Coverage: paths, methods, authentication and success status codes are complete.
    Response bodies are described for every endpoint except a handful that assemble
    them dynamically — each of those says so in its own response description.
    Query parameters and request bodies are not described yet.
  license:
    name: MIT
    url: https://github.com/ufna/birdman/blob/main/LICENSE
servers:
  - url: '{master}'
    description: Your own birdman master. Every install has its own address; there is no hosted service.
    variables:
      master:
        default: http://127.0.0.1:8100
        description: Base URL of the master — the default of the deploy/ compose stack.
paths:
...
  /v1/alerts/active:
    get:
      operationId: alertsActive
      summary: Currently firing alerts
      description: Requires an API key with the "readonly" scope (or "admin").
      security:
        - apiKey: []
      responses:
        4XX:
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        5XX:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.activeAlertsResp'
...
  /v1/alerts/mutes:
    get:
      operationId: listAlertMutes
      summary: List alert mutes
      description: Requires an API key with the "readonly" scope (or "admin").
      security:
        - apiKey: []
      responses:
        4XX:
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        5XX:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.alertMutesResp'
    post:
      operationId: createAlertMute
      summary: Mute an alert
      description: Requires an API key with the "admin" scope (or "admin").
      security:
        - apiKey: []
      responses:
        4XX:
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        5XX:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        "201":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.alertMuteResp'
...
  /v1/alerts/rules:
    get:
      operationId: alertRules
      summary: List alert rules
      description: Requires an API key with the "readonly" scope (or "admin").
      security:
        - apiKey: []
      responses:
        4XX:
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        5XX:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.alertRulesResp'
...
  /v1/environments/{project}/{name}/usage:
    parameters:
      - name: project
        in: path
        required: true
        schema:
          type: string
      - name: name
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: environmentUsage
      summary: What an environment currently holds (pre-delete inventory)
      description: Requires an API key with the "readonly" scope (or "admin").
      security:
        - apiKey: []
      responses:
        4XX:
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        5XX:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.environmentUsageResp'
...
  /v1/mcp:
    get:
      operationId: mcpGet
      summary: 'MCP endpoint: open the server-to-client stream'
      description: Credentials travel in the request body, not in an Authorization header.
      security: []
      responses:
        4XX:
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        5XX:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        "200":
          description: Success. Body is not described by this contract (an upstream proxy response, an event stream, a non-JSON payload or no body at all).
    post:
      operationId: mcpPost
      summary: 'MCP endpoint: call a tool (Model Context Protocol, streamable HTTP)'
      description: Credentials travel in the request body, not in an Authorization header.
      security: []
      responses:
        4XX:
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        5XX:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        "200":
          description: Success. Body is not described by this contract (an upstream proxy response, an event stream, a non-JSON payload or no body at all).
    delete:
      operationId: mcpDelete
      summary: 'MCP endpoint: end the session'
      description: Credentials travel in the request body, not in an Authorization header.
      security: []
      responses:
        4XX:
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        5XX:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        "200":
          description: Success. Body is not described by this contract (an upstream proxy response, an event stream, a non-JSON payload or no body at all).
...
  /v1/nodes:
    get:
      operationId: listNodes
      summary: List fleet nodes
      description: Requires an API key with the "readonly" scope (or "admin").
      security:
        - apiKey: []
      responses:
        4XX:
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        5XX:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.nodesResp'
    post:
      operationId: createNode
      summary: Enroll a node into the fleet
      description: Requires an API key with the "admin" scope (or "admin").
      security:
        - apiKey: []
      responses:
        4XX:
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        5XX:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        "201":
          description: Success. This endpoint's response body is not described yet (it is assembled dynamically); see docs/specs/master.md §6.
...
  /v1/openapi.yaml:
    get:
      operationId: openAPI
      summary: This API's own OpenAPI 3.1 contract, generated from the route table
      description: 'Public by design: no API key required.'
      security: []
      responses:
        4XX:
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        5XX:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        "200":
          description: Success. Body is not described by this contract (an upstream proxy response, an event stream, a non-JSON payload or no body at all).
  /v1/projects:
    get:
      operationId: listProjects
      summary: List projects
      description: Requires an API key with the "readonly" scope (or "admin").
      security:
        - apiKey: []
      responses:
        4XX:
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        5XX:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.projectsResp'
    post:
      operationId: createProject
      summary: Create a project; fails if the slug is taken
      description: Requires an API key with the "admin" scope (or "admin").
      security:
        - apiKey: []
      responses:
        4XX:
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        5XX:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        "201":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.projectResp'
...
  /v1/promote:
    post:
      operationId: promote
      summary: Promote a version into another environment
      description: Requires an API key with the "deploy" scope (or "admin").
      security:
        - apiKey: []
      responses:
        4XX:
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        5XX:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        "200":
          description: Success. This endpoint's response body is not described yet (it is assembled dynamically); see docs/specs/master.md §6.
...
  /v1/servers:
    get:
      operationId: listServers
      summary: List dedicated servers and their states
      description: Requires an API key with the "readonly" scope (or "admin").
      security:
        - apiKey: []
      responses:
        4XX:
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        5XX:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.serversResp'
...
  /v1/versions:
    get:
      operationId: listVersions
      summary: List versions with their fleet state
      description: Requires an API key with the "readonly" scope (or "admin").
      security:
        - apiKey: []
      responses:
        4XX:
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        5XX:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.versionsResp'
    post:
      operationId: createVersion
      summary: Register a build version
      description: Requires an API key with the "deploy" scope (or "admin").
      security:
        - apiKey: []
      responses:
        4XX:
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        5XX:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        "201":
          description: Success. This endpoint's response body is not described yet (it is assembled dynamically); see docs/specs/master.md §6.
components:
  schemas:
...
    httpapi.activeAlertsResp:
      type: object
      properties:
        alerts:
          type:
            - array
            - "null"
          items:
            $ref: '#/components/schemas/httpapi.activeAlert'
      required:
        - alerts
...
    httpapi.alertMuteResp:
      type: object
      properties:
        mute:
          $ref: '#/components/schemas/store.AlertMute'
      required:
        - mute
    httpapi.alertMutesResp:
      type: object
      properties:
        mutes:
          type:
            - array
            - "null"
          items:
            $ref: '#/components/schemas/store.AlertMute'
      required:
        - mutes
...
    httpapi.alertRulesResp:
      type: object
      properties:
        rules:
          type:
            - array
            - "null"
          items:
            $ref: '#/components/schemas/httpapi.alertRule'
      required:
        - rules
    httpapi.apiError:
      type: object
      properties:
        detail:
          type: string
        error:
          type: string
      required:
        - detail
        - error
...
    httpapi.environmentUsageResp:
      type: object
      properties:
        usage:
          $ref: '#/components/schemas/store.EnvironmentUsage'
      required:
        - usage
...
    httpapi.nodesResp:
      type: object
      properties:
        nodes:
          type:
            - array
            - "null"
          items:
            $ref: '#/components/schemas/store.Node'
      required:
        - nodes
...
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: 'birdman API key (bmk_…) in `Authorization: Bearer`. A key carries scopes (admin, deploy, matchmaking, allocate, readonly) and may be bound to one (project, environment) pair, outside of which it can reach nothing.'
    sessionCookie:
      type: apiKey
      in: cookie
      name: birdman_session
      description: Browser session cookie issued by POST /v1/session; the panel uses it instead of carrying an API key in every request.